% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireDataset.R
\name{acquireDataset}
\alias{acquireDataset}
\alias{releaseDataset}
\title{Acquire the Arrow Dataset}
\usage{
acquireDataset(path, ...)

releaseDataset(path)
}
\arguments{
\item{path}{String specifying a path to a Parquet data directory or file.}

\item{...}{Further arguments to be passed to \code{\link[arrow]{open_dataset}}.}
}
\value{
For \code{acquireDataset}, an Arrow Dataset identical to that returned by \code{\link[arrow]{open_dataset}}.

For \code{releaseDataset}, any existing Dataset for the \code{path} is cleared from cache, and \code{NULL} is invisibly returned.
If \code{path=NULL}, all cached Datasets are removed.
}
\description{
Acquire a (possibly cached) Arrow Dataset created from Parquet data.
}
\details{
\code{acquireDataset} will cache the Dataset object in the current R session to avoid repeated initialization.
This improves efficiency for repeated calls, e.g., when creating a \linkS4class{DataFrame} with multiple columns from the same Parquet data path.
The cached Dataset for any given \code{path} can be deleted by calling \code{releaseDataset} for the same \code{path}.
}
\examples{
# Mocking up a single parquet file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

acquireDataset(tf)
acquireDataset(tf) # just re-uses the cache
releaseDataset(tf) # clears the cache

# Mocking up a parquet data diretory:
td <- tempfile()
on.exit(unlink(td), add = TRUE)
arrow::write_dataset(mtcars, td, format = "parquet", partitioning = "cyl")

acquireDataset(td)
acquireDataset(td) # just re-uses the cache
releaseDataset(td) # clears the cache
}
\author{
Aaron Lun, Patrick Aboyoun
}
