% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireHandle.R
\name{acquireHandle}
\alias{acquireHandle}
\alias{releaseHandle}
\title{Acquire the Parquet file handle}
\usage{
acquireHandle(path)

releaseHandle(path)
}
\arguments{
\item{path}{String containing a path to a Parquet file.}
}
\value{
For \code{acquireHandle}, an Arrow Table identical to that returned by \code{\link{read_parquet}} with \code{as_data_frame=FALSE}.

For \code{releaseHandle}, any existing Table for the \code{path} is cleared from cache, and \code{NULL} is invisibly returned.
If \code{path=NULL}, all cached handles are removed.
}
\description{
Acquire a (possibly cached) handle to a Parquet file, given its path.
}
\details{
\code{acquireHandle} will cache the file handle in the current R session to avoid repeated initialization.
This improves efficiency for repeated calls, e.g., when creating a \linkS4class{DataFrame} with multiple columns from the same Parquet file.
The cached handle for any given \code{path} can be removed from the cache by calling \code{releaseHandle}.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

acquireHandle(tf)
acquireHandle(tf) # just re-uses the cache
releaseHandle(tf) # clears the cache
}
\author{
Aaron Lun
}
