% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireTable.R
\name{acquireTable}
\alias{acquireTable}
\alias{releaseTable}
\title{Acquire the Arrow Table}
\usage{
acquireTable(path)

releaseTable(path)
}
\arguments{
\item{path}{String containing a path to a Parquet file.}
}
\value{
For \code{acquireTable}, an Arrow Table identical to that returned by \code{\link{read_parquet}} with \code{as_data_frame=FALSE}.

For \code{releaseTable}, any existing Table for the \code{path} is cleared from cache, and \code{NULL} is invisibly returned.
If \code{path=NULL}, all cached Tables are removed.
}
\description{
Acquire a (possibly cached) Arrow Table representing a Parquet file given its path.
}
\details{
\code{acquireTable} will cache the Table object in the current R session to avoid repeated initialization.
This improves efficiency for repeated calls, e.g., when creating a \linkS4class{DataFrame} with multiple columns from the same Parquet file.
The cached Table for any given \code{path} can be deleted by calling \code{releaseTable} for the same \code{path}.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

acquireTable(tf)
acquireTable(tf) # just re-uses the cache
releaseTable(tf) # clears the cache
}
\author{
Aaron Lun
}
