% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_query.R
\name{arrow_query}
\alias{arrow_query}
\alias{arrow_query,ParquetArray-method}
\alias{arrow_query,ParquetArraySeed-method}
\alias{arrow_query,ParquetColumn-method}
\alias{arrow_query,ParquetDataFrame-method}
\alias{arrow_query,ParquetMatrix-method}
\title{Arrow Query Accessor}
\usage{
arrow_query(x, ...)
}
\arguments{
\item{x}{An object to get or set the arrow query value.}

\item{...}{Additional arguments, for use in specific methods.}
}
\description{
Get or set the arrow query value contained in an object.
}
\examples{
arrow_query
showMethods("arrow_query")

}
\author{
Patrick Aboyoun
}
