% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keynames.R
\name{keynames}
\alias{keynames}
\alias{nkey}
\alias{keynames,ParquetFactTable-method}
\alias{nkey,ParquetFactTable-method}
\title{Key Names and Key Count}
\usage{
keynames(x, ...)
}
\arguments{
\item{x}{An object to get the key names or key count.}

\item{...}{Additional arguments, for use in specific methods.}
}
\description{
Get the key names or key count of an object.
}
\examples{
keynames
showMethods("keynames")

nkey
showMethods("nkey")

}
\author{
Patrick Aboyoun
}
