% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetArray.R
\name{ParquetArray}
\alias{ParquetArray}
\alias{ParquetArray-class}
\alias{[,ParquetArray,ANY,ANY,ANY-method}
\alias{aperm,ParquetArray-method}
\alias{t,ParquetArray-method}
\alias{Arith,ParquetArray,ParquetArray-method}
\alias{Arith,ParquetArray,numeric-method}
\alias{Arith,numeric,ParquetArray-method}
\alias{Compare,ParquetArray,ParquetArray-method}
\alias{Compare,ParquetArray,numeric-method}
\alias{Compare,numeric,ParquetArray-method}
\alias{Logic,ParquetArray,ParquetArray-method}
\alias{Math,ParquetArray-method}
\title{Parquet datasets as DelayedArray objects}
\usage{
ParquetArray(data, key, value, type = NULL, ...)
}
\arguments{
\item{data}{Either a string containing the path to the Parquet data,
or an \code{arrow_dplyr_query} object.}

\item{key}{Either a character vector or a named list of character vectors
containing the names of the columns in the Parquet data that specify
the primary key of the array.}

\item{value}{String containing the name of the column in the Parquet data
that specifies the value of the array.}

\item{type}{String specifying the type of the Parquet data values;
one of \code{"logical"}, \code{"integer"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, this is determined by inspecting
the data.}

\item{...}{Further arguments to be passed to
\code{\link[arrow]{open_dataset}}.}
}
\description{
The ParquetArray class is a \link[DelayedArray]{DelayedArray} subclass
for representing and operating on a Parquet dataset.

All the operations available for \link[DelayedArray]{DelayedArray}
objects work on ParquetArray objects.
}
\examples{
# Create a data.frame from the Titanic data
df <- do.call(expand.grid, c(dimnames(Titanic), stringsAsFactors = FALSE))
df$fate <- Titanic[as.matrix(df)]

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqarray <- ParquetArray(tf, key = c("Class", "Sex", "Age", "Survived"), value = "fate")

}
\seealso{
\code{\link{ParquetArraySeed}},
\code{\link[DelayedArray]{DelayedArray}}
}
\author{
Patrick Aboyoun
}
