% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetArraySeed.R
\name{ParquetArraySeed}
\alias{ParquetArraySeed}
\alias{ParquetArraySeed-class}
\alias{[,ParquetArraySeed,ANY,ANY,ANY-method}
\alias{aperm,ParquetArraySeed-method}
\alias{DelayedArray,ParquetArraySeed-method}
\alias{dim,ParquetArraySeed-method}
\alias{dimnames,ParquetArraySeed-method}
\alias{extract_array,ParquetArraySeed-method}
\alias{t,ParquetArraySeed-method}
\alias{type,ParquetArraySeed-method}
\title{ParquetArraySeed objects}
\usage{
ParquetArraySeed(data, dimensions, value, type = NULL, ...)
}
\arguments{
\item{data}{Either a string containing the path to the Parquet data,
or an existing ParquetArraySeed object.}

\item{dimensions}{Either a character vector or a named list of character
vectors containing the names of the columns in the Parquet data that specify
the dimensions of the array.}

\item{value}{String containing the name of the column in the Parquet data
that specifies the value of the array.}

\item{type}{String specifying the type of the Parquet data values;
one of \code{"logical"}, \code{"integer"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, this is determined by inspecting
the data.}

\item{...}{Further arguments to be passed to
\code{\link[arrow]{open_dataset}}.}
}
\description{
ParquetArraySeed is a low-level helper class for representing a
pointer to a Parquet dataset.

Note that a ParquetArraySeed object is not intended to be used directly.
Most end users will typically create and manipulate a higher-level
\link{ParquetArray} object instead. See \code{?\link{ParquetArray}} for
more information.
}
\examples{
# Create a data.frame from the Titanic data
df <- do.call(expand.grid, c(dimnames(Titanic), stringsAsFactors = FALSE))
df$fate <- as.integer(Titanic[as.matrix(df)])

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqaseed <- ParquetArraySeed(tf, dimensions = c("Class", "Sex", "Age", "Survived"), value = "fate")

}
\seealso{
\code{\link{ParquetArray}},
\code{\link[S4Arrays]{Array}}
}
\author{
Patrick Aboyoun
}
