% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetColumn.R
\name{ParquetColumn}
\alias{ParquetColumn}
\alias{ParquetColumn-class}
\alias{as.vector,ParquetColumn-method}
\alias{extractROWS,ParquetColumn,ANY-method}
\alias{head,ParquetColumn-method}
\alias{length,ParquetColumn-method}
\alias{names,ParquetColumn-method}
\alias{show,ParquetColumn-method}
\alias{showAsCell,ParquetColumn-method}
\alias{tail,ParquetColumn-method}
\alias{Ops,ParquetColumn,ParquetColumn-method}
\alias{Ops,ParquetColumn,vector-method}
\alias{Ops,vector,ParquetColumn-method}
\alias{Math,ParquetColumn-method}
\title{ParquetColumn objects}
\description{
ParquetColumn objects
}
\author{
Patrick Aboyoun
}
