% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetColumnSeed.R
\name{ParquetColumnSeed}
\alias{ParquetColumnSeed}
\alias{ParquetColumnSeed-class}
\alias{dim,ParquetColumnSeed-method}
\alias{type,ParquetColumnSeed-method}
\alias{extract_array,ParquetColumnSeed-method}
\alias{ParquetColumnVector-class}
\alias{ParquetColumnVector}
\title{Delayed column of a Parquet file}
\usage{
ParquetColumnSeed(path, column, type = NULL, length = NULL)

ParquetColumnVector(x, ...)
}
\arguments{
\item{path}{String containing a path to a Parquet file.}

\item{column}{String containing the name of the column inside the file.}

\item{type}{String specifying the type of the data.
If \code{NULL}, this is determined by inspecting the file.
Users may specify this to avoid a look-up, or to coerce the output into a different type.}

\item{length}{Integer containing the number of rows.
If \code{NULL}, this is determined by inspecting the file.
This should only be supplied for efficiency purposes, to avoid a file look-up on construction.}

\item{x}{Either a string containing the path to a Parquet file (to be used as \code{path}),
or an existing ParquetColumnSeed object.}

\item{...}{Further arguments to be passed to the \code{ParquetColumnSeed} constructor.}
}
\value{
For \code{ParquetColumnSeed}, a ParquetColumnSeed is returned, obviously.

For \code{ParquetColumnVector}, a ParquetColumnVector is returned.
}
\description{
Represent a column of a Parquet file as a 1-dimensional DelayedArray.
This allows us to use Parquet data inside \linkS4class{DataFrame}s without loading them into memory.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

# Creating a vector:
ParquetColumnVector(tf, column="gear")

# This happily lives inside DataFrames:
collected <- list()
for (x in colnames(mtcars)) {
    collected[[x]] <- ParquetColumnVector(tf, column=x)
}
DataFrame(collected)

}
\author{
Aaron Lun
}
