% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetDataFrame.R
\name{ParquetDataFrame}
\alias{ParquetDataFrame}
\alias{ParquetDataFrame-class}
\alias{nrow,ParquetDataFrame-method}
\alias{ncol,ParquetDataFrame-method}
\alias{length,ParquetDataFrame-method}
\alias{rownames,ParquetDataFrame-method}
\alias{names,ParquetDataFrame-method}
\alias{rownames<-,ParquetDataFrame-method}
\alias{names<-,ParquetDataFrame-method}
\alias{extractROWS,ParquetDataFrame-method}
\alias{extractCOLS,ParquetDataFrame-method}
\alias{[[,ParquetDataFrame-method}
\alias{replaceROWS,ParquetDataFrame-method}
\alias{replaceCOLS,ParquetDataFrame-method}
\alias{normalizeSingleBracketReplacementValue,ParquetDataFrame-method}
\alias{[[<-,ParquetDataFrame-method}
\alias{cbind,ParquetDataFrame-method}
\alias{cbind.ParquetDataFrame}
\alias{as.data.frame,ParquetDataFrame-method}
\alias{coerce,ParquetDataFrame,DFrame-method}
\title{Parquet-backed DataFrame}
\usage{
ParquetDataFrame(path, columns = NULL, nrows = NULL)
}
\arguments{
\item{path}{String containing a path to a Parquet file.}

\item{columns}{Character vector containing the names of columns in a Parquet file.
If \code{NULL}, this is determined from \code{path}.}

\item{nrows}{Integer scalar specifying the number of rows in a Parquet file.
If \code{NULL}, this is determined from \code{path}.}
}
\value{
A ParquetDataFrame.
}
\description{
Create a Parquet-backed \linkS4class{DataFrame}, where the data are kept on disk until requested.
}
\details{
The ParquetDataFrame is equivalent to a \linkS4class{DataFrame} of \linkS4class{ParquetColumnVector} objects,
and is useful for indicating that the in-memory representation is consistent with the underlying Parquet file.
Operations on a ParquetDataFrame may return another ParquetDataFrame if the operation does not add any information that is inconsistent with the file contents,
e.g., slicing by column will return a ParquetDataFrame as the contents of each column are unchanged.
Other operations will cause the ParquetDataFrame to collapse to a regular \linkS4class{DFrame} of \linkS4class{ParquetColumnVector} objects;
these are still file-backed but lack the guarantee of file consistency.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

# Creating the data frame:
ParquetDataFrame(tf)

}
\author{
Aaron Lun
}
