% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetMatrix.R
\name{ParquetMatrix}
\alias{ParquetMatrix}
\alias{ParquetMatrix-class}
\alias{[,ParquetMatrix,ANY,ANY,ANY-method}
\title{Parquet datasets as DelayedMatrix objects}
\usage{
ParquetMatrix(data, row, col, value, key = c(row, col), type = NULL, ...)
}
\arguments{
\item{data}{Either a string containing the path to the Parquet data,
or an existing ParquetArraySeed object.}

\item{row}{Either a character vector or a named list of character vectors
containing the names of the columns in the Parquet data that specify the
rows of the matrix.}

\item{col}{Either a character vector or a named list of character vectors
containing the names of the columns in the Parquet data that specify the
columns of the matrix.}

\item{value}{String containing the name of the column in the Parquet data
that specifies the value of the matrix.}

\item{key}{Either a character vector or a named list of character vectors
containing the names of the columns in the Parquet data that specify the
rows and columns of the matrix.}

\item{type}{String specifying the type of the Parquet data values;
one of \code{"logical"}, \code{"integer"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, this is determined by inspecting
the data.}

\item{...}{Further arguments to be passed to
\code{\link[arrow]{open_dataset}}.}
}
\description{
The ParquetMatrix class is a \link[DelayedArray]{DelayedMatrix} subclass
for representing and operating on a Parquet dataset.

All the operations available for \link[DelayedArray]{DelayedMatrix}
objects work on ParquetMatrix objects.
}
\examples{
# Create a data.frame from a matrix
df <- data.frame(
  rowname = rep(rownames(state.x77), times = ncol(state.x77)),
  colname = rep(colnames(state.x77), each = nrow(state.x77)),
  value = as.vector(state.x77)
)

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqmat <- ParquetMatrix(tf, row = "rowname", col = "colname", value = "value")

}
\author{
Patrick Aboyoun
}
