% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query,ParquetColumnSeed-method}
\alias{query,ParquetColumnVector-method}
\alias{query,ParquetDataFrame-method}
\title{Query Accessor}
\usage{
query(x, ...)
}
\arguments{
\item{x}{An object to get or set the query value.}

\item{...}{Additional arguments, for use in specific methods.}
}
\description{
Get or set the query value contained in an object.
}
\examples{
query
showMethods("query")

}
\author{
Patrick Aboyoun
}
