% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireDataset.R
\name{acquireDataset}
\alias{acquireDataset}
\alias{releaseDataset}
\title{Acquire the Arrow Dataset}
\usage{
acquireDataset(path)

releaseDataset(path)
}
\arguments{
\item{path}{String containing a path to a Parquet file.}
}
\value{
For \code{acquireDataset}, an Arrow Dataset identical to that returned by \code{\link{read_parquet}} with \code{as_data_frame=FALSE}.

For \code{releaseDataset}, any existing Dataset for the \code{path} is cleared from cache, and \code{NULL} is invisibly returned.
If \code{path=NULL}, all cached Datasets are removed.
}
\description{
Acquire a (possibly cached) Arrow Dataset representing a Parquet file given its path.
}
\details{
\code{acquireDataset} will cache the Dataset object in the current R session to avoid repeated initialization.
This improves efficiency for repeated calls, e.g., when creating a \linkS4class{DataFrame} with multiple columns from the same Parquet file.
The cached Dataset for any given \code{path} can be deleted by calling \code{releaseDataset} for the same \code{path}.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

acquireDataset(tf)
acquireDataset(tf) # just re-uses the cache
releaseDataset(tf) # clears the cache
}
\author{
Aaron Lun
}
