% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetColumnSeed.R
\name{ParquetColumnSeed}
\alias{ParquetColumnSeed}
\alias{ParquetColumnSeed-class}
\alias{dim,ParquetColumnSeed-method}
\alias{type,ParquetColumnSeed-method}
\alias{Arith,ParquetColumnSeed,ParquetColumnSeed-method}
\alias{Arith,ParquetColumnSeed,numeric-method}
\alias{Arith,numeric,ParquetColumnSeed-method}
\alias{Compare,ParquetColumnSeed,ParquetColumnSeed-method}
\alias{Compare,ParquetColumnSeed,vector-method}
\alias{Compare,vector,ParquetColumnSeed-method}
\alias{Logic,ParquetColumnSeed,ParquetColumnSeed-method}
\alias{Math,ParquetColumnSeed-method}
\alias{DelayedArray,ParquetColumnSeed-method}
\alias{extract_array,ParquetColumnSeed-method}
\alias{ParquetColumnVector-class}
\alias{show,ParquetColumnVector-method}
\alias{Ops,ParquetColumnVector,ParquetColumnVector-method}
\alias{Ops,ParquetColumnVector,numeric-method}
\alias{Ops,numeric,ParquetColumnVector-method}
\alias{Math,ParquetColumnVector-method}
\alias{ParquetColumnVector}
\title{Column of a Parquet file}
\usage{
ParquetColumnSeed(path, column, type = NULL, length = NULL, ...)

ParquetColumnVector(x, column, type = NULL, length = NULL, ...)
}
\arguments{
\item{path}{String specifying a path to a Parquet data directory or file.}

\item{column}{String containing the name of the column in the data.}

\item{type}{String specifying the type of the data.
If \code{NULL}, this is determined by inspecting the data.
Users may specify this to avoid a look-up, or to coerce the output into a different type.}

\item{length}{Integer containing the number of rows.
If \code{NULL}, this is determined by inspecting the data.
This should only be supplied for efficiency purposes, to avoid a file look-up on construction.}

\item{...}{Further arguments to be passed to \code{\link[arrow]{open_dataset}}.}

\item{x}{Either a string containing the path to the Parquet data (to be used as \code{path}),
or an existing ParquetColumnSeed object.}
}
\value{
For \code{ParquetColumnSeed}, a ParquetColumnSeed is returned, obviously.

For \code{ParquetColumnVector}, a ParquetColumnVector is returned.
}
\description{
Represent a column of a Parquet file as a 1-dimensional \linkS4class{DelayedArray}.
This allows us to use Parquet data inside \linkS4class{DataFrame}s without loading them into memory.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

# Creating a vector:
ParquetColumnVector(tf, column="gear")

# This happily lives inside DataFrames:
collected <- list()
for (x in colnames(mtcars)) {
    collected[[x]] <- ParquetColumnVector(tf, column=x)
}
DataFrame(collected)

}
\author{
Aaron Lun, Patrick Aboyoun
}
