% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetDataFrame.R
\name{ParquetDataFrame}
\alias{ParquetDataFrame}
\alias{ParquetDataFrame-class}
\alias{makeNakedCharacterMatrixForDisplay,ParquetDataFrame-method}
\alias{nrow,ParquetDataFrame-method}
\alias{ncol,ParquetDataFrame-method}
\alias{length,ParquetDataFrame-method}
\alias{rownames,ParquetDataFrame-method}
\alias{names,ParquetDataFrame-method}
\alias{rownames<-,ParquetDataFrame-method}
\alias{names<-,ParquetDataFrame-method}
\alias{extractROWS,ParquetDataFrame,ANY-method}
\alias{head,ParquetDataFrame-method}
\alias{tail,ParquetDataFrame-method}
\alias{extractCOLS,ParquetDataFrame-method}
\alias{[[,ParquetDataFrame-method}
\alias{replaceROWS,ParquetDataFrame-method}
\alias{replaceCOLS,ParquetDataFrame-method}
\alias{normalizeSingleBracketReplacementValue,ParquetDataFrame-method}
\alias{[[<-,ParquetDataFrame-method}
\alias{bindROWS,ParquetDataFrame-method}
\alias{cbind,ParquetDataFrame-method}
\alias{cbind.ParquetDataFrame}
\alias{as.data.frame,ParquetDataFrame-method}
\title{Parquet-backed DataFrame}
\arguments{
\item{data}{Either a string containing the path to the Parquet data,
or an \code{arrow_dplyr_query} object.}

\item{key}{Either a character vector or a named list of character vectors
containing the names of the columns in the Parquet data that specify
the primary key of the array.}

\item{...}{Further arguments to be passed to \code{\link[arrow]{open_dataset}}.}
}
\value{
A ParquetDataFrame where each column is a \linkS4class{ParquetColumn}.
}
\description{
Create a Parquet-backed \linkS4class{DataFrame}, where the data are kept on disk until requested.
}
\details{
The ParquetDataFrame is essentially just a \linkS4class{DataFrame} of \linkS4class{ParquetColumn} objects.
It is primarily useful for indicating that the in-memory representation is consistent with the underlying Parquet data
(e.g., no delayed filter/mutate operations have been applied, no data has been added from other files).
Thus, users can specialize code paths for a ParquetDataFrame to operate directly on the underlying Parquet data.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(cbind(model = rownames(mtcars), mtcars), tf)

# Creating our Parquet-backed data frame:
df <- ParquetDataFrame(tf, key = "model")
df

# Extraction yields a ParquetColumn:
df$carb

# Slicing ParquetDataFrame objects:
df[,1:5]
df[1:5,]

# Combining by ParquetDataFrame and ParquetColumn objects:
combined <- cbind(df, df)
class(combined)
combined2 <- cbind(df, some_new_name=df[,1])
class(combined2)

}
\author{
Aaron Lun, Patrick Aboyoun
}
