% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetDataFrame.R
\name{ParquetDataFrame}
\alias{ParquetDataFrame}
\alias{ParquetDataFrame-class}
\alias{show,ParquetDataFrame-method}
\alias{nrow,ParquetDataFrame-method}
\alias{ncol,ParquetDataFrame-method}
\alias{length,ParquetDataFrame-method}
\alias{rownames,ParquetDataFrame-method}
\alias{names,ParquetDataFrame-method}
\alias{rownames<-,ParquetDataFrame-method}
\alias{names<-,ParquetDataFrame-method}
\alias{extractROWS,ParquetDataFrame,ANY-method}
\alias{head,ParquetDataFrame-method}
\alias{tail,ParquetDataFrame-method}
\alias{extractCOLS,ParquetDataFrame-method}
\alias{[[,ParquetDataFrame-method}
\alias{replaceROWS,ParquetDataFrame-method}
\alias{replaceCOLS,ParquetDataFrame-method}
\alias{normalizeSingleBracketReplacementValue,ParquetDataFrame-method}
\alias{[[<-,ParquetDataFrame-method}
\alias{cbind,ParquetDataFrame-method}
\alias{cbind.ParquetDataFrame}
\alias{as.data.frame,ParquetDataFrame-method}
\alias{coerce,ParquetDataFrame,DFrame-method}
\title{Parquet-backed DataFrame}
\usage{
ParquetDataFrame(path, ...)
}
\arguments{
\item{path}{String specifying a path to a Parquet data directory or file.}

\item{...}{Further arguments to be passed to \code{\link[arrow]{open_dataset}}.}
}
\value{
A ParquetDataFrame where each column is a \linkS4class{ParquetColumnVector}.
}
\description{
Create a Parquet-backed \linkS4class{DataFrame}, where the data are kept on disk until requested.
}
\details{
The ParquetDataFrame is essentially just a \linkS4class{DataFrame} of \linkS4class{ParquetColumnVector} objects.
It is primarily useful for indicating that the in-memory representation is consistent with the underlying Parquet data
(e.g., no delayed filter/mutate operations have been applied, no data has been added from other files).
Thus, users can specialize code paths for a ParquetDataFrame to operate directly on the underlying Parquet data.

In that vein, operations on a ParquetDataFrame may return another ParquetDataFrame if the operation does not introduce inconsistencies with the file-backed data.
For example, slicing or combining by column will return a ParquetDataFrame as the contents of the retained columns are unchanged.
In other cases, the ParquetDataFrame will collapse to a regular \linkS4class{DFrame} of \linkS4class{ParquetColumnVector} objects before applying the operation;
these are still file-backed but lack the guarantee of file consistency.
}
\examples{
# Mocking up a file:
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(mtcars, tf)

# Creating our Parquet-backed data frame:
df <- ParquetDataFrame(tf)
df

# Extraction yields a ParquetColumnVector:
df$carb

# Some operations preserve the ParquetDataFrame:
df[,1:5]
combined <- cbind(df, df)
class(combined)

# ... but most operations collapse to a regular DFrame:
df[1:5,]
combined2 <- cbind(df, some_new_name=df[,1])
class(combined2)

}
\author{
Aaron Lun, Patrick Aboyoun
}
